requirements.txt
flask~=3.0.3
python-dotenv~=1.0.1
gunicorn
Flask-Mail
bleach~=6.2.0
email-validator~=2.2.0
flask-recaptcha
flask-limiter~=3.9.2
redis
celery

requests~=2.32.3